+++ CHANGELOG.md
@@ -13,4 +13,9 @@
 11. Tested List All Employees by Salary showing Role, Salary and Department
 12. Created View All Employees query, called from server.js and formatted using console.table
 13. Added queries for View All Departments and View All Roles
-14. Added query to View All Employees in Specific Department
+14. Added query to View All Employees by specific Department or specific Manager
+15. Added code to add new Department
+16. Added code to add new Role
+17. Added code to add new Employee
+18. Added code to update Employees Role
+19. Added code to delete an Employee